Fixed 64-bit issues
git-svn-id: svn://svn.code.sf.net/p/shareaza/code/trunk@8137 3231ed4b-2112-0410-84e0-8c3944a40730

+++ shareaza/GProfile.cpp
@@ -97,7 +97,7 @@ void CGProfile::CreateBT()
 	// Convert Gnutella GUID (128 bits) to BitTorrent GUID (160 bits)
 	Hashes::BtGuid tmp_bt;
 	CopyMemory( &tmp_bt[0], &((Hashes::Guid)oGUID)[0], ((Hashes::Guid)oGUID).byteCount );
-	for ( int nByte = ((Hashes::Guid)oGUID).byteCount ; nByte < tmp_bt.byteCount ; nByte++ )
+	for ( size_t nByte = ((Hashes::Guid)oGUID).byteCount ; nByte < tmp_bt.byteCount ; nByte++ )
 		tmp_bt[ nByte ] = GetRandomNum( 0ui8, _UI8_MAX );
 	
 	VERIFY( tmp_bt.validate() );
